1. Added comments 2. Refactored some method 3. UiViewController extension added for tap gesture 4. Attributed text utility method added
